chunair/kicad-footprint-generator-master/requirements.txt
pyyaml
future
pycodestyle
pytest
